+++ packages/manufacturing/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @happyvertical/smrt-manufacturing
 
+## 0.40.53
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.53
+  - @happyvertical/smrt-inventory@0.40.53
+  - @happyvertical/smrt-tenancy@0.40.53
+
+## 0.40.52
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.52
+  - @happyvertical/smrt-inventory@0.40.52
+  - @happyvertical/smrt-tenancy@0.40.52
+
+## 0.40.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.51
+  - @happyvertical/smrt-inventory@0.40.51
+  - @happyvertical/smrt-tenancy@0.40.51
+
+## 0.40.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.50
+  - @happyvertical/smrt-inventory@0.40.50
+  - @happyvertical/smrt-tenancy@0.40.50
+
 ## 0.40.49
 
 ### Patch Changes