+++ packages/smrt-playground/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @happyvertical/smrt-playground
 
+## 0.40.53
+
+### Patch Changes
+
+- @happyvertical/smrt-ui@0.40.53
+
+## 0.40.52
+
+### Patch Changes
+
+- @happyvertical/smrt-ui@0.40.52
+
+## 0.40.51
+
+### Patch Changes
+
+- @happyvertical/smrt-ui@0.40.51
+
+## 0.40.50
+
+### Patch Changes
+
+- @happyvertical/smrt-ui@0.40.50
+
 ## 0.40.49
 
 ### Patch Changes