+++ packages/smrt-svelte/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @happyvertical/smrt-svelte
 
+## 0.40.53
+
+### Patch Changes
+
+- @happyvertical/smrt-languages@0.40.53
+- @happyvertical/smrt-types@0.40.53
+- @happyvertical/smrt-ui@0.40.53
+- @happyvertical/smrt-web@0.40.53
+
+## 0.40.52
+
+### Patch Changes
+
+- @happyvertical/smrt-languages@0.40.52
+- @happyvertical/smrt-types@0.40.52
+- @happyvertical/smrt-ui@0.40.52
+- @happyvertical/smrt-web@0.40.52
+
+## 0.40.51
+
+### Patch Changes
+
+- @happyvertical/smrt-languages@0.40.51
+- @happyvertical/smrt-types@0.40.51
+- @happyvertical/smrt-ui@0.40.51
+- @happyvertical/smrt-web@0.40.51
+
+## 0.40.50
+
+### Patch Changes
+
+- @happyvertical/smrt-languages@0.40.50
+- @happyvertical/smrt-types@0.40.50
+- @happyvertical/smrt-ui@0.40.50
+- @happyvertical/smrt-web@0.40.50
+
 ## 0.40.49
 
 ### Patch Changes